Create issue template to redirect people to isaacs repo
This repository was created for a specific purpose, which was
fulfilled back in 2016.  However people are still submitting issues
about GitHub to the repository's issue tracker.  This effectively
fragments and weakens the GitHub community, because

    https://github.com/isaacs/github

was already established in 2013 for the same purpose, and has acquired
very significant momentum since then.  It is more efficient for the
community to pool its efforts in a single place, so make an issue
template which explains the current situation and encourages people
to submit issues to isaacs instead of here.

This partially addresses #207.  When combined with a PR auto-closer
bot as suggested in

    https://github.com/isaacs/github/issues/1191

the combination could be considered as a full solution for #207.